Factor out issues/routes.ts.
